Switch to euclidean distance matching for cycleDimensions
Fixes matching issues when the desktop modifies the resize request
(eg. for gnome-panel).

Introduces a potential issue where the user can no longer be certain
that a given number of keypresses will take an un-tiled window to a
given state. This could easily be an issue since it removes the ability
to operate by muscle memory and a solution is being considered.